# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API de la plateforme QUI PERD (défis de match, escrow, paiements Mobile Money).",
        "title": "QUI PERD API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/utilisateurs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "utilisateurs"
                ],
                "summary": "Liste des comptes (admin) — paginée, 10 par page, plus récents d'abord",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page (défaut 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Éléments par page (1..100, défaut 10)",
                        "name": "taille",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Pseudo ou e-mail (recherche partielle)",
                        "name": "recherche",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "actif | suspendu | en_attente | supprime",
                        "name": "statut",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.Page-auth_Utilisateur"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "utilisateurs"
                ],
                "summary": "Créer un compte joueur (admin) — mêmes règles que l'inscription, portefeuille créé, aucune session",
                "parameters": [
                    {
                        "description": "Compte à créer",
                        "name": "corps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utilisateurs.EntreeCreation"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/auth.Utilisateur"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/utilisateurs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "utilisateurs"
                ],
                "summary": "Détail d'un compte (admin) : utilisateur + portefeuille { soldeDisponible, soldeBloque }",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utilisateurs.UtilisateurDetail"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "utilisateurs"
                ],
                "summary": "Suppression logique d'un compte (admin) — refusée (409) si de l'argent ou un match est en jeu",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "utilisateurs"
                ],
                "summary": "Mettre à jour un compte (propriétaire : profil ; admin : tout, dont email, statut, motDePasse)",
                "parameters": [
                    {
                        "description": "Champs à modifier (tous facultatifs)",
                        "name": "corps",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/utilisateurs.EntreeModification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.Utilisateur"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "auth.Utilisateur": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "dateModification": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nomUtilisateur": {
                    "type": "string"
                },
                "pays": {
                    "type": "string"
                },
                "photoProfil": {
                    "type": "string"
                },
                "statut": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "utilisateurs.EntreeCreation": {
            "type": "object",
            "required": [
                "email",
                "motDePasse",
                "nomUtilisateur"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "motDePasse": {
                    "type": "string",
                    "minLength": 6
                },
                "nomUtilisateur": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "pays": {
                    "type": "string"
                },
                "statut": {
                    "type": "string",
                    "enum": [
                        "actif",
                        "suspendu",
                        "en_attente"
                    ]
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "utilisateurs.EntreeModification": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "motDePasse": {
                    "type": "string",
                    "minLength": 6
                },
                "nomUtilisateur": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 3
                },
                "pays": {
                    "type": "string"
                },
                "photoProfil": {
                    "type": "string"
                },
                "statut": {
                    "type": "string",
                    "enum": [
                        "actif",
                        "suspendu",
                        "en_attente"
                    ]
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
        "utilisateurs.PortefeuilleResume": {
            "type": "object",
            "properties": {
                "soldeBloque": {
                    "type": "number"
                },
                "soldeDisponible": {
                    "type": "number"
                }
            }
        },
        "utilisateurs.UtilisateurDetail": {
            "type": "object",
            "properties": {
                "dateCreation": {
                    "type": "string"
                },
                "dateModification": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "nomUtilisateur": {
                    "type": "string"
                },
                "pays": {
                    "type": "string"
                },
                "photoProfil": {
                    "type": "string"
                },
                "portefeuille": {
                    "$ref": "#/definitions/utilisateurs.PortefeuilleResume"
                },
                "statut": {
                    "type": "string"
                },
                "telephone": {
                    "type": "string"
                }
            }
        },
...
        "utils.Page-auth_Utilisateur": {
            "type": "object",
            "properties": {
                "elements": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/auth.Utilisateur"
                    }
                },
                "page": {
                    "type": "integer"
                },
                "pages": {
                    "type": "integer"
                },
                "taille": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}